refactor: rename dump_state
